+++ src/redux/actions.js
@@ -4,3 +4,10 @@ export function updateUser (user) {
         payload: user
     };
 }
+
+export function changeOnlineTutors(tutors) {
+  return {
+    type: 'CHANGE_ONLINE_TUTORS',
+    payload: tutors
+  };
+}
